Add Blazor initialization files and Bootstrap CSS
Introduce a set of foundational files required for initializing and running a Blazor application. These files (_Host.cshtml, _Imports.razor, App.razor, and appsettings files) are standard files which are needed to correctly bootstrap a Blazor application. In order to improve the UI of our application, the Bootstrap CSS is also added.