Add User Retrieval Feature
Added a new feature that allows user retrieval by ID. This involved the addition of new methods and classes related to User retrieval in the Application and Domain layer. Included in the changes are UserRepository.cs, IUserRepository.cs, UserController.cs, and several new classes in the UserFeatures folder. The changes in UserController.cs enable the API to accept and respond to HTTP GET requests for user data retrieval. The feature was added to meet the need for retrieving user details in the system.