Refactored IUserRepository and mapper registration
In this commit, FindById method was removed from the IUserRepository, and switched to using the GetByIdAsync method from the IBaseRepository. This change builds on the existing functionality of the IBaseRepository, reducing redundancy.

Furthermore, the method for registering mappers was also updated in ServiceExtensions.cs. This was done to streamline the mapper registration process, replacing the individually added singleton mappers with an automated registration method using reflection. This change improves code organization and promotes scalability as new mappers can be automatically registered.