Implemented License Key Validation Service
Added a new background service to periodically check for license key validity. This service uses a singleton of `LicenseKeyService` that implements `ILicenseKeyService` to verify the status of a license key and logs the result to the console. Changes also included an update to `ServiceExtensions` to register the new service, and a relocation of `IntervalHostedService` from Application to Persistence. The motivation behind these modifications is to establish periodic checks of license validity, which is an crucial part of maintaining software's functionality and compliance.