Remove FetchData feature and refactor Counter
Removed the unused FetchData feature to clean up the application. This included deleting the FetchData.razor file and its reference in the NavMenu.razor. Also, refactored the Counter feature by injecting UserClientService into Counter.razor, instead of creating a new instance in the code. This change is to promote Dependency Injection principles and improve testability. Lastly, added HttpClient and UserClientService to services in Program.cs for DI.