Integrate MediatR for domain event dispatch
MediatR library is introduced to enable domain events dispatch in the base repository. This is done to ensure that these events will be processed after a commit operation by mapping them to event handlers. The commits include moving SSHManagement and UserManagement features to separate directories and renaming their classes. Mediator is added to SSHCommander and UserRepository constructors, along with BaseRepository, making it possible to publish events across the application. The Refactoring of application layer services to accommodate new usage of domain events is also done. New domain event CreateUserDomainEvent has been added and mapped to CreateUserDomainEventHandler. Unit of works now will be responsible in calling the domain events. SSHCommanderRepository is also added to the service collection in Persistence project's ServiceExtensions. This allows the application to handle domain events asynchronously and enhances code maintainability. Lastly, necessary classes for MongoDbContext seeding are included but commented out for future implementation requirements. This commit is a major step towards utilizing CQRS pattern and enhancing the overall architecture of the application.