Add WPF application to SwitchCommander and implement basic user UI
This commit introduces a new WPF application into the SwitchCommander solution. The application's MainWindow has basic UI for user information input - email, name1, and name2 with a "Create" button to submit. A MainWindowViewModel is created to bind with the frontend inputs and the command action of the "Create" button. DelegateCommand.cs is added to implement the ICommand interface. The solution file is also updated to include the newly added WPF application project. UserClientService.cs is generated to handle API calls. This setup is to provide a basic GUI to interact with the User model for the next steps of development.