Add AddRangeAsync method to handle bulk insert
Added a new method, AddRangeAsync, in the BaseRepository and IBaseRepository to handle bulk insert operations. This method takes a collection of entities and a CancellationToken as input and inserts all the entities into the collection. This was necessary for optimizing the performance of data seed operation, which previously performed inserts one by one. Now in MongoDbContextSeed, AddRangeAsync is used to perform bulk insert.