Refactor transaction handling methods
Removed `UnitOfWork` and `IUnitOfWork` and moved the transaction handling methods `Commit` and `Rollback` to `BaseRepository` and `IBaseRepository`. The refactoring was needed to simplify the code base and place transaction handling closer to the database operations. Because MongoDB doesn't support explicit transactions for single document operations, this change improves the adaptability of the code for different database systems.