Update .github/workflows/dotnet.yml to include develop branch
The .github/workflows/dotnet.yml file was updated to trigger workflow runs not just on pushes or pull requests to the 'master' branch, but also to the 'develop' branch. This change was deemed necessary to automate testing in the separate 'develop' branch, ensuring code stability before changes are merged into 'master'.