Update .NET badge label in README.md
The .NET GitHub actions badge in the README.md file has been updated from ".NET" to ".NET build". This has been done to make it more descriptive and clear to anyone seeing the badge, indicating that it represents the status of the .NET building process within the GitHub Actions workflows.