Refactor folder structure for more clarity
Change the structure of the folders to make it cleaner and easier to understand. This includes changing the naming of the folders as well as the files in the folders. The 'SSHManagement' and 'UserManagement' folders were renamed to 'SSH' and 'User' respectively, and all the files within these folders were updated to reflect these changes.

This reorganization helps improve project structure and enhances the readability and manageability of the code.