Update SSHCommand model and related classes
Updated SSHCommand model, replacing Hash property with Name and Description properties. Added new SShCommandConfiguration model, and corresponding repository, and methods to handle these models. Added validation and update methods for SSHCommand configurations. Updated existing repository methods to return boolean status on update or delete operations.

This is done to improve the usability and flexibility of our software. The Name and Description properties make it easier for users to identify and understand the purpose of each SSHCommand. SShCommandConfiguration allows for customizable validation on SSHCommands.