Add SSH.NET package and implement SSHNetService
Added the SSH.NET package to the project in order to manage SSH connections directly from our application. Implemented the SSHNetService which currently supports opening an SSH connection and running a basic command. The 'CloseConnection' method is yet to be implemented. SSH related configurations and services folders have been removed as they are no longer needed.