Refactor SSH related features for cleaner code structure
Various changes are made to improve the clarity and organisation of the SSH feature's code structure. Most notably in SSHController, IValidator instances for different request types are explicitly named to indicate their uses and responsibilities for enhanced readability. A new server creation endpoint is also added. The ISSHServerRepository interface and its implementation are rearranged and simplified, moving command execution responsibility away. Further changes include the removal of unused using directives and better organisation of the services addition to the IServiceCollection. Removed execution from SSHServerRepository as responsibility for command execution was moved elsewhere,