Refactor SSH feature classes to appropriate namespaces
Organized the SSH-related feature classes into their appropriate namespaces based on the functionality they provide.
UpdateSSHCommand was deleted due to unnecessary duplication and was replaced with a new class. UpdateSSHServerHandler.cs class was created for better management of SSH updates and ensuring that the id of a guid type. Similarly, classes related to other operations like Create, Delete, and Read have also been renamed and moved to their corresponding namespaces. This structure will make it easier to manage and understand these classes in the future.