Refactor SSH feature handling and introduce updated operations
This refactor was done to simplify SSH feature handling by removing unnecessary classes and files while improving the code organization. The 'UpdateSSHCommandHandler.cs' was deleted as it was not being used and the SSH feature handling was shifted from the 'UserController.cs' to 'SSHServerController.cs' to better align with its functionality. The 'UpdateAsync' was changed to 'ReplaceAsync' in the 'UpdateSSHCommandsConfigurationHandler.cs' and 'BaseRepository.cs' to better reflect its functionality. Additionally, the 'UpdateSSHServerPasswordHandler', 'UpdateSSHServerPasswordValidator', and 'UpdateSSHServerPasswordRequest' classes were added for handling changes to the SSH server password.