"Implement Read all functionality across application"
The application has undergone considerable changes in this commit to support the 'Read All' operations.

Many elements including handlers, mappers, repositories were added or modified. The Handlers, 'ReadAllSSHConfigurationHandler' and 'ReadAllSSHServerHandler' are introduced to handle the requests and responses for SSH configuration and server respectively. Repositories were updated to refer to the 'ReadAllAsync' function which retrieves all entities. In 'SSHCommandService', 'SSHExecuteCommandService', 'SSHConfigurationService', 'SSHServerService', 'UserService', the NotImplementedException is replaced with actual functionality calls. Other services and classes also had similar modifications for better consistency and ease of use.

The renaming of 'SwitchCommandApi' to be inside `Apis` folder was needed for better directory structuring.

These changes are important to allow users to retrieve all data from different parts of the application in a consistent and reliable manner.