Add SSH Export functionality and update database settings
Add new functionality for exporting SSH commands, servers, sequences, and configs in json format. The export type can be selected ("command", "server", "sequence", "config", or "blob"). If an invalid export type is selected, a new ExportTypeException is thrown.

Also, updates were made to the MongoDB and Hangfire ConnectionStrings in the appsettings.json files for easier local development. The SSHCommandSequenceMongoRepository was renamed to SshSequenceMongoRepository for consistency, and respective references in other files were updated.

Lastly, two new controller files for SSHImport and SSHExport were added. The exact functionality of these is not yet implemented. Similarly, a new UpdateSSHImportJsonHandler class was created without implemented functionality.