Add SSH import feature in SwitchCommander Application
A new feature for SSH import functionality has been added in the SwitchCommander application. This involves the setting up of a new folder for SSH import under Features and implementing the related controller. The controller uses MediatR for handling request-response with in-process messaging and extends the BaseController for shared functionalities. This new feature is important for supporting import tasks related to SSH.