Update sequence & command types, add NotMapped attribute
Changed the type of SshCommands in SSHSequence.cs from List to IEnumerable for better performance and increased flexibility. This allows for deferred execution. Also, in CreateSSHSequenceHandler.cs, the Commands type in the CreateSSHCommandSequenceRequest record was updated to SSHCommand from SSHCommandSequenceRequestDto for better readability and correctness.

In BaseEntity.cs, added the [NotMapped] attribute to the DomainEvents property to prevent it from being mapped to a database column during the database creation, as this property is not intended to be persisted in the DB.