Add XML comments for Swagger and enable XML docs generation
This commit enables the generation of XML documentation files in the WebAPI and WPF projects and configures Swagger to use these XML comments. XML docs were also added for one method in SSHCommandController for testing purposes. This was implemented in order to provide more detailed API documentation in Swagger UI.